api 수정

+++ src/routes/api.js
@@ -1,10 +1,11 @@
 const express = require('express');
 const router = express.Router();
 const db = require('../config/db');
+const auth = require('./common/auth');
 
 /**
  * 오늘 요일의 전체 웹툰보기 API
- * localhost:3000/webtoons/{day}?page={page}&sort={sortType}
+ * localhost:3000/api/webtoons/{day}?sort={sortType}
  */
 router.get('/list/:day', function(request, response, next){
     const day = request.params.day;
@@ -25,7 +26,7 @@ router.get('/list/:day', function(request, response, next){
 
 /**
  * 플랫폼별 웹툰 보기 API
- * localhost:3000/webtoons/{day}/{platform}?page={page}&sort={sortType}
+ * localhost:3000/api/webtoons/{day}/{platform}?sort={sortType}
  */
  router.get('/list/:day/:platform', function(request, response, next) {
     const { day, platform } = request.params;
@@ -46,7 +47,7 @@ router.get('/list/:day', function(request, response, next){
 
 /**
  * 웹툰명, 작가명으로 웹툰 검색 API
- * localhost:3000/webtoons/search?name={search_name}page={page}&sort={sortType}
+ * localhost:3000/api/webtoons/search?name={search_name}?sort={sortType}
  */
 router.get('/search', function(request, response, next){
     let { name, sort } = request.query;
@@ -84,6 +85,11 @@ function filterQueryBySortType(sort, condition){
 
 // 사용자 즐겨찾기 보기 API
 router.get('/favorites', function(request, response){
+    const user = auth.getLoginUser(request, response);
+    if(user === undefined){
+        alert("로그인이 필요한 서비스입니다.");
+        return;
+    }
     const sql = `SELECT * FROM webtoon as w INNER JOIN favorites as f on w.id = f.webtoon_id WHERE f.user_id = ?`;
     db.query(sql, user.id, function(error, favorites){
         if(error) {
@@ -92,17 +98,22 @@ router.get('/favorites', function(request, response){
         }
         response.send(favorites);
     });
-  })
+})
   
   //즐겨찾기 추가하기 API
 router.post('/favorites', function(request, response){
+    const user = auth.getLoginUser(request, response);
+    if(user === undefined){
+        alert("로그인이 필요한 서비스입니다.");
+        return;
+    }
     const webtoon_id = request.body.webtoon_id;
     const is_favorite = request.body.is_favorite;
     let sql = '';
-    if(is_favorite){
-    sql += `INSERT INTO favorites (user_id, webtoon_id) VALUES (?, ?)`;
+    if(is_favorite == 'false'){
+        sql += `INSERT INTO favorites (user_id, webtoon_id) VALUES (?, ?)`;
     } else {
-    sql += `DELETE FROM favorites WHERE user_id = ? AND webtoon_id = ?`;
+        sql += `DELETE FROM favorites WHERE user_id = ? AND webtoon_id = ?`;
     }
 
     db.query(sql, [user.id, webtoon_id], function(error, result){
@@ -112,6 +123,7 @@ router.post('/favorites', function(request, response){
         }
         response.send(result);
     });
+    
 });
 
 module.exports = router;
